refactor: CP utils (#83)
* add cp_utils


* fix


* fix


* fix


* fix


* fix


* resolve conflict


* Fix bug


---------
