+++ nemo_automodel/distributed/parallelizer.py
@@ -12,11 +12,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import contextlib
 import importlib
 import signal
 from functools import lru_cache
-from typing import Dict, Generator, List, Optional, Set, Union
+from typing import Dict, List, Optional, Union
 
 import torch
 from torch import Tensor, nn
@@ -361,71 +360,3 @@ def _destroy_dist_connection() -> None:
         torch.distributed.destroy_process_group()
     signal.signal(signal.SIGINT, signal.SIG_DFL)
 
-
-# based on https://github.com/pytorch/torchtitan/blob/main/torchtitan/distributed/utils.py#L113
-def create_context_parallel_ctx(
-    cp_mesh: DeviceMesh,
-    cp_buffers: List[torch.Tensor],
-    cp_seq_dims: List[int],
-    cp_no_restore_buffers: Set[torch.Tensor],
-    cp_rotate_method: str,
-):
-    """
-    Create a context parallel context.
-
-    Args:
-        cp_mesh (DeviceMesh): The device mesh for context parallel.
-        cp_buffers (List[torch.Tensor]): The buffers for context parallel.
-        cp_seq_dims (List[int]): The sequence dimensions for context parallel.
-        cp_no_restore_buffers (Set[torch.Tensor]): The no restore buffers for context parallel.
-        cp_rotate_method (str): The rotation method for context parallel, such as "allgather" or "addtoall".
-    """
-    from torch.distributed.tensor.experimental import context_parallel
-
-    # TODO: uncomment this when torch.distributed.tensor.experimental._attention.set_rotate_method is available
-    # from torch.distributed.tensor.experimental._attention import set_rotate_method
-    # set_rotate_method(cp_rotate_method)
-    return context_parallel(
-        cp_mesh,
-        buffers=cp_buffers,
-        buffer_seq_dims=cp_seq_dims,
-        no_restore_buffers=cp_no_restore_buffers,
-    )
-
-
-# based on https://github.com/pytorch/torchtitan/blob/main/torchtitan/distributed/utils.py#L138
-def get_train_context(enable_loss_parallel: bool, enable_compiled_autograd: bool):
-    """
-    Create a train context.
-
-    Args:
-        enable_loss_parallel (bool): Whether to enable loss parallelism.
-        enable_compiled_autograd (bool): Whether to enable compiled autograd.
-    """
-
-    @contextlib.contextmanager
-    def context(cp_context: Optional[Generator[None, None, None]] = None):
-        with contextlib.ExitStack() as stack:
-            if enable_loss_parallel:
-                stack.enter_context(torch.distributed.tensor.parallel.loss_parallel())
-
-            if enable_compiled_autograd:
-                stack.enter_context(
-                    torch._dynamo.utils.maybe_enable_compiled_autograd(True)
-                )
-
-            if cp_context is not None:
-                from torch.nn.attention import SDPBackend, sdpa_kernel
-
-                # currently we only support these two SDP backends.
-                # SDPBackend.MATH is not currently compatible with DTensor
-                stack.enter_context(
-                    sdpa_kernel(
-                        [SDPBackend.FLASH_ATTENTION, SDPBackend.EFFICIENT_ATTENTION]
-                    )
-                )
-                stack.enter_context(cp_context)
-
-            yield
-
-    return context
